Split all the validators
